+++ src/_init.lua
@@ -96,7 +96,6 @@ end
 do
   local Confirmer = Addon:GetModule("Confirmer")
   local L = Addon:GetModule("Locale")
-  local ListItemParser = Addon:GetModule("ListItemParser")
   local Seller = Addon:GetModule("Seller")
 
   --- Returns `true` with a reason string if a critical process is active.
@@ -104,7 +103,6 @@ do
   --- @return string? reason
   function Addon:IsBusy()
     if Seller:IsBusy() then return true, L.IS_BUSY_SELLING_ITEMS end
-    if ListItemParser:IsBusy() then return true, L.IS_BUSY_UPDATING_LISTS end
     if Confirmer:IsBusy() then return true, L.IS_BUSY_CONFIRMING_ITEMS end
     return false
   end